+++ src/contexts/theme/ThemeAMB.tsx
@@ -1,0 +1,7 @@
+import {JSX} from 'react';
+
+import('./themeAMB.css');
+
+export default function ThemeAMB(): JSX.Element {
+	return <></>;
+}
