+++ src/contexts/theme/ThemeDAA.tsx
@@ -1,0 +1,7 @@
+import {JSX} from 'react';
+
+import('./themeDAA.css');
+
+export default function ThemeDAA(): JSX.Element {
+	return <></>;
+}
