+++ src/contexts/theme/ThemeTTT.tsx
@@ -1,0 +1,7 @@
+import {JSX} from 'react';
+
+import('./themeTTT.css');
+
+export default function ThemeTTT(): JSX.Element {
+	return <></>;
+}
