+++ src/contexts/theme/themeAMB.css
@@ -1,0 +1,21 @@
+body, main {
+	background-color: light-dark(var(--mantine-color-gray-1), var(--mantine-color-dark-7));
+}
+
+*::-webkit-scrollbar {
+	width: 8px;
+}
+
+*::-webkit-scrollbar-thumb {
+	background-color: var(--mantine-color-gray-5);
+	border-radius: 6px;
+}
+
+@supports (-moz-appearance:none) {
+	* {
+		/*The following attributes are currently supported only by Firefox. Webkit browsers are designed by the ::-webkit-scrollbar
+		So that nothing is broken in potential future support, these values are set only for Firefox.*/
+		scrollbar-color: var(--mantine-color-gray-5) transparent;
+		scrollbar-width: thin;
+	}
+}
